Rename shared lib to clash_lib
