+++ clash-server/Cargo.toml
@@ -1,26 +1,18 @@
 [package]
-name = "client"
+name = "clash-server"
 version.workspace = true
 edition.workspace = true
 publish = false
-
-[features]
-console = []
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 anyhow.workspace = true
-bfbb = { version = "0.2.1", features = ["game-interface"] }
+bfbb.workspace = true
 clash_lib.workspace = true
+tokio.workspace = true
+thiserror.workspace = true
 env_logger.workspace = true
 log.workspace = true
-thiserror.workspace = true
-tokio.workspace = true
 
-eframe = "0.19"
-image = "0.24"
-spin_sleep = "1"
-
-[target.'cfg(windows)'.build-dependencies]
-winres = "0.1"
+rand = "0.8"
